requirements.txt
tensorflow>=2.13
opencv-python>=4.8.0
mediapipe>=0.10.0
albumentations>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
numpy>=1.24.0
